requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
idna==3.6
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.4
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.13.3
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
